colocate event handler with worker
